requirements.txt
numpy==1.20.3
Pillow==9.0.1
pytorch_lightning==1.5.10
sentence_transformers==2.0.0
torch==1.10.2
torchmetrics==0.7.2
tqdm==4.62.2
transformers==4.10.0
